hiding Delete and Edit button if not looking at their own profile added in /profile page

+++ frontend/src/routes/profile/+page.svelte
@@ -1,7 +1,7 @@
 <script>
   import { goto } from "$app/navigation";
   import Button from "$lib/components/Button.svelte";
-
+  
   export let data;
 </script>
 
@@ -24,24 +24,26 @@
           <Button name="Read" onClick={() => {
             goto(`/article?${new URLSearchParams({ id: _id, email })}`);
           }}/>
-          <Button name="Edit" onClick={() => {
-            goto('/edit-article?' + new URLSearchParams({ id: _id }));
-          }}/>
-          <Button name="Delete" onClick={() => {
-            fetch(`http://localhost:3000/users/article/${title}`, {
-              method: "DELETE", 
-              headers: {
-                'Authorization': `Bearer ${sessionStorage.getItem('jwt')}`,
-                'Content-type': 'application/json; charset=UTF-8'
-              },
-            });
+          {#if data.isMyProfile}
+            <Button name="Edit" onClick={() => {
+              goto('/edit-article?' + new URLSearchParams({ id: _id }));
+            }}/>
+            <Button name="Delete" onClick={() => {
+              fetch(`http://localhost:3000/users/article/${title}`, {
+                method: "DELETE", 
+                headers: {
+                  'Authorization': `Bearer ${sessionStorage.getItem('jwt')}`,
+                  'Content-type': 'application/json; charset=UTF-8'
+                },
+              });
 
-            fetch(`http://localhost:3000/articles/${_id}`, {
-              method: "DELETE", 
-            });
+              fetch(`http://localhost:3000/articles/${_id}`, {
+                method: "DELETE", 
+              });
 
-            data.articles = data.articles.filter(article => article._id !== _id);
-          }}/>
+              data.articles = data.articles.filter(article => article._id !== _id);
+            }}/>
+          {/if}
         </div>
       </div>
       {/each}
